requirements.txt
# Core Django Framework
Django==5.2.6
python-dotenv~=1.0

# REST Framework and CORS Handling
djangorestframework
django-cors-headers

# PDF Generation for Receipts
reportlab~=4.0

# Development Tools
django-extensions